Rename models/ to realmModels/ and use babel aliases
