+++ babel.config.js
@@ -13,6 +13,7 @@ module.exports = {
         // for some reason, this doesn't seem to work for models, so I'm leaving that directory out
         navigation: "./src/navigation",
         providers: "./src/providers",
+        realmModels: "./src/realmModels",
         sharedHelpers: "./src/sharedHelpers",
         sharedHooks: "./src/sharedHooks",
         styles: "./src/styles"
