+++ src/components/ObsDetails/ObsDetails.js
@@ -28,15 +28,15 @@ import {
 } from "react-native";
 import { ActivityIndicator, Button as IconButton } from "react-native-paper";
 import IconMaterial from "react-native-vector-icons/MaterialIcons";
+import Observation from "realmModels/Observation";
+import Taxon from "realmModels/Taxon";
+import User from "realmModels/User";
 import { formatObsListTime } from "sharedHelpers/dateAndTime";
 import useApiToken from "sharedHooks/useApiToken";
 import useRemoteObservation from "sharedHooks/useRemoteObservation";
 import colors from "styles/colors";
 import { imageStyles, textStyles, viewStyles } from "styles/obsDetails/obsDetails";
 
-import Observation from "../../models/Observation";
-import Taxon from "../../models/Taxon";
-import User from "../../models/User";
 import ActivityTab from "./ActivityTab";
 import DataTab from "./DataTab";
 import checkCamelAndSnakeCase from "./helpers/checkCamelAndSnakeCase";
