+++ src/providers/contexts.js
@@ -1,8 +1,8 @@
 // @flow
 import { createRealmContext } from "@realm/react";
 import { createContext } from "react";
-
-import realmConfig from "../models/index";
+// eslint-disable-next-line import/extensions
+import realmConfig from "realmModels/index";
 
 const ExploreContext: Object = createContext<Function>( );
 const ObsEditContext: Object = createContext<Function>( );
