+++ src/sharedHooks/useCVSuggestions.js
@@ -3,8 +3,7 @@
 import { getJWTToken } from "components/LoginSignUp/AuthenticationService";
 import inatjs, { FileUpload } from "inaturalistjs";
 import { useEffect, useState } from "react";
-
-import Photo from "../models/Photo";
+import Photo from "realmModels/Photo";
 
 const TAXON_FIELDS = {
   name: true,
