+++ src/sharedHooks/useRemoteObservation.js
@@ -4,9 +4,8 @@ import NetInfo from "@react-native-community/netinfo";
 import { getUsername } from "components/LoginSignUp/AuthenticationService";
 import inatjs from "inaturalistjs";
 import { useEffect, useRef, useState } from "react";
-
-import Observation from "../models/Observation";
-import User from "../models/User";
+import Observation from "realmModels/Observation";
+import User from "realmModels/User";
 
 const useRemoteObservation = ( observation: Object, refetch: boolean ): Object => {
   const [remoteObservation, setRemoteObservation] = useState( null );
