+++ tests/unit/components/Observations/ObsCard.test.js
@@ -10,8 +10,8 @@ const testObservation = factory( "LocalObservation" );
 const qualityGradeText = t( "RG" );
 
 // this probably isn't the right approach, but it does allow the test to pass
-jest.mock( "../../../../src/models/index", ( ) => {
-  const originalModule = jest.requireActual( "../../../../src/models/index" );
+jest.mock( "../../../../src/realmModels/index", ( ) => {
+  const originalModule = jest.requireActual( "../../../../src/realmModels/index" );
 
   // Mock the default export and named export 'foo'
   return {
